+++ ddtrace/contrib/fastapi/__init__.py
@@ -57,8 +57,12 @@ required_modules = ["fastapi"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.fastapi.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.fastapi.patch import get_version
+        from ..internal.fastapi.patch import patch
+        from ..internal.fastapi.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
