+++ ddtrace/contrib/flask/__init__.py
@@ -104,10 +104,9 @@ required_modules = ["flask"]
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
         # DEV: We do this so we can `@mock.patch('ddtrace.contrib.flask._patch.<func>')` in tests
-        from . import patch as _patch
-
-        patch = _patch.patch
-        unpatch = _patch.unpatch
-        get_version = _patch.get_version
+        from . import patch as _  # noqa: F401, I001
+        from ..internal.flask.patch import patch
+        from ..internal.flask.patch import unpatch
+        from ..internal.flask.patch import get_version
 
         __all__ = ["patch", "unpatch", "get_version"]
